Allow applying all manifests at once
We currently only apply the `hypershift-operator.yaml` manifests which
causes changes in other files to not get rolled out. Applying the whole
folder doesn't work straight away, because it contains the kustomize
documents.

Kustomize doesn't support referencing manifests in a parent dir, so we
move all rendered manifests in a new manifests subfolder.